storage-module/requirements.txt
kafka-python-ng
psycopg2-binary
redis